% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_model_selections.R
\name{collate_model_selections}
\alias{collate_model_selections}
\title{Bayesian linear regression with Stan}
\usage{
collate_model_selections(loomfile, loo_results, attr_name)
}
\arguments{
\item{loomfile}{Numeric vector of UMI counts.}

\item{loo_results}{Numeric vector of cell sizes (total UMI counts per cell).}

\item{attr_name}{Name of the .}
}
\value{
A list of `stanfit` returned by four models: Poisson, Negative-Binomial, Zero-Inflated Poisson, & Zero-Inflated Negative-Binomial
}
\description{
Bayesian linear regression with Stan
}
