% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_bestfit_genes.R
\name{find_bestfit_genes}
\alias{find_bestfit_genes}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
find_bestfit_genes(fit_file, model2check, margin = 2, loomfile = NULL,
  attr_name = NULL, verbose = FALSE)
}
\arguments{
\item{fit_file}{A file name that contains model fitting result reside}

\item{margin}{A multiplier for standard deviation (SD) in leave-one-out ELPD for calling models}

\item{loomfile}{A expression quantity file (loom format)}

\item{attr_name}{Name of the row attribute in loomfile for storing best model calls}

\item{verbose}{Whether to print out overall model calls}
}
\value{
best_model_calls Genes that Best fits to a given model are also stored in the input loomfile
}
\description{
Bayesian model selection for scRNA-seq count data
}
