% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_model_best.R
\name{find_model_best}
\alias{find_model_best}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
find_model_best(fit_list, model2find, margin = 2, loomfile = NULL,
  attr_name = NULL, verbose = FALSE)
}
\arguments{
\item{fit_list}{A list of scRATE model fit results}

\item{model2find}{Model of interest on finding representative genes}

\item{margin}{A multiplier for standard deviation (SD) in leave-one-out ELPD for calling models}

\item{loomfile}{A expression quantity file (loom format)}

\item{attr_name}{Name of the row attribute in loomfile for storing best model calls}

\item{verbose}{Whether to print out overall model calls}
}
\value{
best_model_calls Genes that Best fits to a given model are also stored in the input loomfile
}
\description{
Bayesian model selection for scRNA-seq count data
}
