% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_count_models.R
\name{fit_count_models}
\alias{fit_count_models}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
fit_count_models(y, exposure, nCores = NULL, seed = NULL,
  model2fit = NULL, brms4zi = FALSE)
}
\arguments{
\item{y}{Numeric vector of UMI counts.}

\item{exposure}{Numeric vector of cell sizes (total UMI counts per cell).}

\item{nCores}{Number of cores.}

\item{seed}{Seed number.}

\item{model2fit}{A specific model to fit (1:P, 2:NB, 3:ZIP, 4:ZINB, NULL:All)}
}
\value{
A list of `stanfit` returned by model(s)
}
\description{
Bayesian model selection for scRNA-seq count data
}
