% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bestmodel_fit.R
\name{get_bestmodel_fit}
\alias{get_bestmodel_fit}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
get_bestmodel_fit(cntfile, nCores = NULL, seed = NULL,
  outfile = NULL)
}
\arguments{
\item{cntfile}{Expression quantity file (RData format: Use 'save' and 'load')}

\item{nCores}{Number of cores}

\item{seed}{Seed number}

\item{outfile}{Output file name to store ELPD_loo results (RDS format)}
}
\value{
A list of model fits (only available when outfile is not specified)
}
\description{
Bayesian model selection for scRNA-seq count data
}
