% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_fit.R
\name{get_model_fit}
\alias{get_model_fit}
\title{Fit regression models for many genes and save it}
\usage{
get_model_fit(cntfile, formula_string = NULL, model2fit = NULL,
  nCores = NULL, seed = NULL, outfile = NULL)
}
\arguments{
\item{cntfile}{Expression quantity file (RData format: Use 'save' and 'load')}

\item{formula_string}{A regression formula to fit the non-zero-inflated counts}

\item{nCores}{Number of cores}

\item{seed}{Seed number}

\item{outfile}{Output file name to store ELPD_loo results (RDS format)}
}
\value{
A list of model fits (only available when outfile is not specified)
}
\description{
Fit regression models for many genes and save it
}
