% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_params.R
\name{get_model_params}
\alias{get_model_params}
\title{An internal function to re-organize stan_fit parameters}
\usage{
get_model_params(model_fit, covariate = NULL)
}
\arguments{
\item{model_fit}{A list of four model fits}

\item{covariate}{A list of covariates if any}
}
\value{
A list of mean parameter values for each model
}
\description{
An internal function to re-organize stan_fit parameters
}
