% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_bestfit.R
\name{is_bestfit}
\alias{is_bestfit}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
is_bestfit(loo_results, model2check, margin = 2)
}
\arguments{
\item{loo_results}{ELPD_loo for models}

\item{model2check}{Model to check whether a gene bestfits to it (Either of 1, 2, 3, or 4)}

\item{margin}{A multiplier for confidence interval}
}
\value{
Whether a given gene best fits to the specified model (boolean)
}
\description{
Bayesian model selection for scRNA-seq count data
}
