% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perform_model_selections.R
\name{perform_model_selection}
\alias{perform_model_selection}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
perform_model_selection(fit_list, margin = 2, loomfile = NULL,
  attr_name = NULL, verbose = FALSE)
}
\arguments{
\item{fit_list}{A list of model fitting results}

\item{margin}{A multiplier for standard deviation (SD) in leave-one-out ELPD for calling models}

\item{loomfile}{A expression quantity file (loom format)}

\item{attr_name}{Name of the row attribute in loomfile for storing best model calls}

\item{verbose}{Whether to print out overall model calls}
}
\value{
modelcall Models called for genes. Stored in the input loomfile if provided.
}
\description{
Bayesian model selection for scRNA-seq count data
}
