% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_job_array.R
\name{prepare_job_array}
\alias{prepare_job_array}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
prepare_job_array(loomfile, num_chunks, outdir, dryrun, layer = NULL,
  covariate = NULL, nCores = NULL, seed = NULL, gene_start = NULL,
  gene_end = NULL, chunk_start = NULL, chunk_end = NULL)
}
\arguments{
\item{loomfile}{Expression quantity file (loom format).}

\item{num_chunks}{Number of chunks.}

\item{outdir}{Name of the folder where results should be stored.}

\item{dryrun}{TRUE if you only want to check the job submission commands.}

\item{layer}{Layer name of the count to use in the loom file.}

\item{covariate}{Name of covariate (A col.attrs name in the input loomfile.)}

\item{nCores}{Number of cores to run stan fitting in parallel}

\item{seed}{Seed number to reproduce randomized results}

\item{gene_start}{Starting gene index to analyze.}

\item{gene_end}{Ending gene index to analyze.}

\item{chunk_start}{Starting chunk index to submit.}

\item{chunk_end}{Ending chunk index to submit.}
}
\value{
... None is returned.
}
\description{
Bayesian model selection for scRNA-seq count data
}
