% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_comparison.R
\name{run_model_comparison}
\alias{run_model_comparison}
\title{Bayesian linear regression with Stan}
\usage{
run_model_comparison(cntfile, outfile, nCores, seed)
}
\arguments{
\item{cntfile}{Numeric vector of UMI counts.}

\item{outfile}{Numeric vector of cell sizes (total UMI counts per cell).}

\item{nCores}{Number of cores.}

\item{seed}{Seed number.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
A list of `stanfit` returned by four models: Poisson, Negative-Binomial, Zero-Inflated Poisson, & Zero-Inflated Negative-Binomial
}
\description{
Bayesian linear regression with Stan
}
