% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRATE-package.R
\docType{package}
\name{scRATE-package}
\alias{scRATE-package}
\alias{scRATE}
\title{The 'scRATE' package.}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org
}
