% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_model.R
\name{select_model}
\alias{select_model}
\title{Select models that best fits the gene data}
\usage{
select_model(loo_results, margin = 2)
}
\arguments{
\item{loo_results}{ELPD_loo for models}

\item{margin}{A multiplier for confidence interval}
}
\value{
selected_model Reports the model scRATE selects (1:P, 2:NB, 3:ZIP, 4:ZINB)
}
\description{
Select models that best fits the gene data
}
