% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_jobs.R
\name{submit_jobs}
\alias{submit_jobs}
\title{Bayesian linear regression with Stan}
\usage{
submit_jobs(loomfile, num_chunks, outdir, dryrun, scriptfile, rfile,
  layer = NULL, g_start = NULL, g_end = NULL, chunk_start = NULL,
  chunk_end = NULL)
}
\arguments{
\item{y}{Numeric vector of UMI counts.}

\item{exposure}{Numeric vector of cell sizes (total UMI counts per cell).}

\item{nCores}{Number of cores.}

\item{seed}{Seed number.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
A list of `stanfit` returned by four models: Poisson, Negative-Binomial, Zero-Inflated Poisson, & Zero-Inflated Negative-Binomial
}
\description{
Bayesian linear regression with Stan
}
