% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{version}
\alias{version}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
version()
}
\value{
... Reports the version of the package:scRATE.
}
\description{
Bayesian model selection for scRNA-seq count data
}
