% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_model_selections.R
\name{collate_model_selections}
\alias{collate_model_selections}
\title{Bayesian linear regression with Stan}
\usage{
collate_model_selections(loomfile, loo_results, attr_name)
}
\arguments{
\item{loomfile}{A expression quantity file (loom format).}

\item{loo_results}{Leave-one-out ELPD result file.}

\item{attr_name}{Name of the row attribute in loomfile for storing best model calls.}
}
\value{
best_model_calls Best model calls are also stored in the input loomfile.
}
\description{
Bayesian linear regression with Stan
}
