% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_results.R
\name{collate_results}
\alias{collate_results}
\title{Collate the results from "run_model_comparison" into a single list (and save in RDS format)}
\usage{
collate_results(loo_dir, globstr = "_scrate*", loo_outfile = NULL)
}
\arguments{
\item{loo_dir}{A folder name in which leave-one-out ELPD result files reside}

\item{globstr}{Search string (wildcard supported) for loo result files (in RDS format)}

\item{loo_outfile}{Name of the file to save collated leave-one-out ELPD results (optional)}
}
\value{
A list of collated results if 'loo_outfile' is not given
}
\description{
Collate the results from "run_model_comparison" into a single list (and save in RDS format)
}
