% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_count_models.R
\name{compare_count_models}
\alias{compare_count_models}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
compare_count_models(model_fit)
}
\arguments{
\item{model_fit}{A list of four model fits}
}
\value{
ELPD_loo result
}
\description{
Bayesian model selection for scRNA-seq count data
}
