% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_count_models.R
\name{fit_count_models}
\alias{fit_count_models}
\title{Fit a gene with regression models}
\usage{
fit_count_models(gexpr, f, nCores = NULL, seed = NULL,
  adapt_delta = 0.8, model2fit = NULL, brms4zi = TRUE)
}
\arguments{
\item{gexpr}{A data frame that containes UMI count (y), log_exposure (exposure), and covariate values}

\item{f}{A regression formula to fit the non-zero-inflated counts}

\item{nCores}{Number of cores}

\item{seed}{Seed number}

\item{adapt_delta}{The target average proposal acceptance probability during Stan’s adaptation period (default:0.8)}

\item{model2fit}{A specific model to fit (1:P, 2:NB, 3:ZIP, 4:ZINB, NULL:All)}

\item{brms4zi}{Whether to run brms for zero-inflated models (Deprecated)}
}
\value{
A list of `stanfit` returned by model(s)
}
\description{
Fit a gene with regression models
}
