% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_count_models.R
\name{fit_count_models}
\alias{fit_count_models}
\title{Bayesian linear regression with Stan}
\usage{
fit_count_models(y, exposure, nCores, seed)
}
\arguments{
\item{y}{Numeric vector of UMI counts.}

\item{exposure}{Numeric vector of cell sizes (total UMI counts per cell).}

\item{nCores}{Number of cores.}

\item{seed}{Seed number.}

\item{...}{Arguments passed to `rstan::sampling` (e.g. iter, chains).}
}
\value{
A list of `stanfit` returned by four models: Poisson, Negative-Binomial, Zero-Inflated Poisson, & Zero-Inflated Negative-Binomial
}
\description{
Bayesian linear regression with Stan
}
