% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_params.R
\name{get_model_params}
\alias{get_model_params}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
get_model_params(model_fit, ctyped = FALSE)
}
\arguments{
\item{model_fit}{A list of four model fits}

\item{ctyped}{Whether model_fit used ctype as covariate}
}
\value{
A list of mean parameter values for each model
}
\description{
Bayesian model selection for scRNA-seq count data
}
