% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_model_best.R
\name{is_model_best}
\alias{is_model_best}
\title{Checks if the gene best fits a certain model significantly better over all the other models}
\usage{
is_model_best(loo_result, model2check, margin = 2)
}
\arguments{
\item{loo_result}{ELPD_loo for models}

\item{model2check}{Model to check whether a gene bestfits to it (Either of 1, 2, 3, or 4)}

\item{margin}{A multiplier for confidence interval}
}
\value{
Whether a given gene best fits to the specified model (boolean)
}
\description{
Checks if the gene best fits a certain model significantly better over all the other models
}
