% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_comparison.R
\name{run_model_comparison}
\alias{run_model_comparison}
\title{Fit regression models for many genes}
\usage{
run_model_comparison(cntfile, formula_string = NULL, nCores = NULL,
  seed = NULL, adapt_delta = 0.8, model2fit = NULL, outfile = NULL)
}
\arguments{
\item{cntfile}{Expression quantity file (RData format: Use 'save' and 'load'). The count matrix is gene x cell.}

\item{formula_string}{A regression formula to fit the non-zero-inflated counts}

\item{nCores}{Number of cores}

\item{seed}{Seed number}

\item{adapt_delta}{The target average proposal acceptance probability during Stan’s adaptation period (default:0.8)}

\item{model2fit}{A specific model to fit (1:P, 2:NB, 3:ZIP, 4:ZINB, NULL:All)}

\item{outfile}{Output file name to store ELPD_loo results (RDS format)}
}
\value{
A list of ELPD_loo results and mean parameters returned by loo::compare
}
\description{
Fit regression models for many genes
}
