% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_model_comparison.R
\name{run_model_comparison}
\alias{run_model_comparison}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
run_model_comparison(cntfile, nCores = NULL, seed = NULL,
  outfile = NULL)
}
\arguments{
\item{cntfile}{Expression quantity file (RData format: Use 'save' and 'load').}

\item{nCores}{Number of cores.}

\item{seed}{Seed number.}

\item{outfile}{Output file name to store ELPD_loo results (RDS format).}
}
\value{
A list of ELPD_loo results returned by loo::compare
}
\description{
Bayesian model selection for scRNA-seq count data
}
