% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_jobs.R
\name{submit_jobs}
\alias{submit_jobs}
\title{Bayesian model selection for scRNA-seq count data}
\usage{
submit_jobs(loomfile, num_chunks, outdir, dryrun, rfile, scriptfile,
  covariates = c(), layer = NULL, nCores = NULL, seed = NULL,
  gene_start = NULL, gene_end = NULL, chunk_start = NULL,
  chunk_end = NULL)
}
\arguments{
\item{loomfile}{Expression quantity file (loom format)}

\item{num_chunks}{Number of chunks}

\item{outdir}{Name of the folder where results should be stored}

\item{dryrun}{TRUE if you only want to check the job submission commands}

\item{rfile}{R script to execute}

\item{scriptfile}{PBS job submission script}

\item{covariates}{Name of covariates (A col.attrs name in the input loomfile)}

\item{layer}{Layer name of the count to use in the loom file}

\item{nCores}{Number of cores to run stan fitting in parallel}

\item{seed}{Seed number to reproduce randomized results}

\item{gene_start}{Starting gene index to analyze}

\item{gene_end}{Ending gene index to analyze}

\item{chunk_start}{Starting chunk index to submit}

\item{chunk_end}{Ending chunk index to submit}
}
\value{
... None is returned
}
\description{
Bayesian model selection for scRNA-seq count data
}
